issue-995: extract browser automation contract/runtime
