+++ crates/tau-browser-automation/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod browser_automation_contract;
+pub mod browser_automation_runtime;
