+++ crates/tau-coding-agent/Cargo.toml
@@ -37,6 +37,7 @@ tau-memory = { path = "../tau-memory" }
 tau-github-issues = { path = "../tau-github-issues" }
 tau-dashboard = { path = "../tau-dashboard" }
 tau-custom-command = { path = "../tau-custom-command" }
+tau-browser-automation = { path = "../tau-browser-automation" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
